Organized Topic type
Roomと同一のライフサイクルを持つ、Roomモデルの一部という判断
